refactor: 파싱할 문자열이 유효한 문자열인지 검증하는 로직을 Calculator에서 StringParser로 이동
