+++ README.md
@@ -14,11 +14,11 @@
 
 ### 요구 사항 쪼개기
 
-- 문자열 구분자는 문자열을 입력받으면 구분자를 기준으로 구분할 수 있다.
-- 문자열 구분자는 기본 구분자(쉼표, 콜론) 외에 커스텀 구분자를 지정할 수 있다.
+- 문자열 파서는 문자열을 입력받으면 구분자를 기준으로 구분할 수 있다.
+- 문자열 파서는 기본 구분자(쉼표, 콜론) 외에 커스텀 구분자를 지정할 수 있다.
   - 커스텀 구분자와 기본 구분자를 함께 사용할 수도 있다. ex) "//;\n1;2,3" => 6
+- 문자열 파서는 숫자 이외의 값 또는 음수를 전달하면 예외를 발생한다.
 
 - 문자열 덧셈 계산기는 빈 문자열 또는 null을 입력받으면 0을 반환한다.
 - 문자열 덧셈 계산기는 문자열을 구분한 결과값을 기준으로 더한 값을 반환한다.
-- 문자열 덧셈 계산기는 숫자 이외의 값 또는 음수를 전달할 수 없다.
-  - 만약 전달되었다면 RuntimeException 예외가 발생한다.
+
