+++ src/main/kotlin/stringpluscalculator/StringPlusCalculator.kt
@@ -7,21 +7,7 @@ object StringPlusCalculator {
             return 0
         }
         val parseResult = StringParser.parser(input)
-        checkValidInput(parseResult)
         return plusElements(parseResult)
-    }
-
-    private fun checkValidInput(parseResult: List<String>) {
-        for (element in parseResult) {
-            val parseNum = element.toIntOrNull() ?: throw RuntimeException()
-            checkNegative(parseNum)
-        }
-    }
-
-    private fun checkNegative(parseNum: Int) {
-        if (parseNum < 0) {
-            throw RuntimeException()
-        }
     }
 
     private fun plusElements(parseResult: List<String>): Int {
